requirements.txt
# Audio
pygame==2.6.1
# Slack
slack-bolt==1.22.0
# Spicetify
websockets==14.1
# Google Calendar
google-api-python-client==2.156.0
google-auth-oauthlib==1.2.1
# Tooling
black==24.10.0
pylint==3.3.2
mypy==1.14.0
# TTS
winrt-Windows.Foundation==2.3.0
winrt-Windows.Media.Core==2.3.0
winrt-Windows.Media.Playback==2.3.0
winrt-Windows.Media.SpeechSynthesis==2.3.0
winrt-Windows.Storage==2.3.0
winrt-Windows.Storage.Streams==2.3.0